requirements.txt
Django==2.1.4
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
